requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-api-response==0.1.0
django-redis==5.4.0
django-environ==0.9.0
django-ckeditor==6.3.2

pyotp==2.9.0
qrcode==8.0
pillow==11.0.0

uvicorn==0.32.0

channels==4.1.0
channels-redis==4.2.0

whitenoise==6.8.2

psycopg2==2.9.4

celery==5.4.0
django-celery-results==2.5.1
django-celery-beat==2.7.0

#paquete para la autenticacion
djoser==2.3.1 

#me parece que es para hacer hash
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0 
djangorestframework-simplejwt==5.3.1

#recordar que en la app de blog esta hecho de otra dorma 
django-ckeditor==6.3.2

django-storages==1.14.4
boto3==1.35.15
botocore==1.35.15
cryptography==41.0.7
rsa==4.9

#Sirve para controlar y limitar la cantidad de login que se puede hacer, es decir, que no se ingrese por fuerza bruta
django-axes==7.0.0

#Generar datos para prueba
Faker==33.0.0 